refactor: moved shortcuts.ts to hooks
