+++ src/hooks/appHooks.ts
@@ -8,7 +8,7 @@ import { useNavigate } from 'react-router-dom'
 import { useGetUserDataMutation, useRefreshTokenMutation } from '../api/auth.api'
 import { EXP_MESSAGE } from '../constants'
 import { ROUTES } from '../routes'
-import { useGoToLoginWithMessage } from '../shortcuts'
+import { useGoToLoginWithMessage } from './shortcuts'
 import { selectCurrentUser, updateCurrentUser } from '../slices/currentUserSlice'
 import type { AppDispatch, RootState } from '../store'
 import { AppCookies, appCookiesNames } from '../types'
