+++ src/hooks/shortcuts.ts
@@ -1,7 +1,7 @@
 import { useDispatch } from "react-redux"
 import { useNavigate } from "react-router-dom"
-import { ROUTES } from "./routes"
-import { setMessage } from "./slices/messageSlice"
+import { ROUTES } from "../routes"
+import { setMessage } from "../slices/messageSlice"
 
 export const useGoToLoginWithMessage = () => {
   const dispatch = useDispatch()
